// nextoops-backend-nodejs/src/subscriptions/enum/subscription-type.enum.ts
export enum SubscriptionType {
  INITIAL_PURCHASE = 'INITIAL_PURCHASE',
  RENEWAL = 'RENEWAL',
  CANCELLATION = 'CANCELLATION',
  UNCANCELLATION = 'UNCANCELLATION',
  NON_RENEWING_PURCHASE = 'NON_RENEWING_PURCHASE',
  SUBSCRIPTION_PAUSED = 'SUBSCRIPTION_PAUSED',
  EXPIRATION = 'EXPIRATION',
  BILLING_ISSUE = 'BILLING_ISSUE',
  PRODUCT_CHANGE = 'PRODUCT_CHANGE',
  TRANSFER = 'TRANSFER',
  SUBSCRIPTION_EXTENDED = 'SUBSCRIPTION_EXTENDED',
  EXPERIMENT_ENROLLMENT = 'EXPERIMENT_ENROLLMENT',
}